Moving common methods to common utils
